only interact with / render actors on curr floor
only render / update actors on current floor

issues:
   - player is not always positioned on stairwell after descending to already
     generated map. consider revising methods like getActorAt() to take the
     current dungeon level in to account.
   - enemies standing on corpses will still be ignored by the PlayerAI. corpses
     should always be moved to the _back_ of the actor list, so that they are
     not detected before living actors
   - ActorFactory should have different methods for different sorts of actors

only be concerned with actors on current floor

send corpses to back of actor list

iterate backward when rendering

actors move to the back of the actor list when dying. if rendering
is done front to back, then dead actors (corpses) will end up being
rendered on top of lives ones on the same tile.